backend_api/requirements.txt
fastapi>=0.100.0,<0.112.0
uvicorn[standard]>=0.20.0,<0.30.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
youtube-transcript-api>=0.6.0,<0.7.0
pytube2>=15.0.12
anthropic>=0.51.0,<1.0.0